Refactor: ver3.0 리스트생성 폴더 생성 및 공용헤더 ver3.0 디자인 적용
