+++ src/app/list/_create/_components/list/SimpleInput.css.ts
